+++ assets/js/src/admin/gutenberg/blocks/cart/save.js
@@ -3,17 +3,6 @@
  * @description Cart block save method.
  */
 
-const saveBlock = (props) => {
-	const { shortcode } = props.attributes;
-
-
-	return (
-		<div
-			className={props.className}
-		>
-			{ shortcode }
-		</div>
-	);
-};
+const saveBlock = props => props.attributes.shortcode;
 
 export default saveBlock;