+++ src/BigCommerce/Plugin.php
@@ -4,7 +4,7 @@
 namespace BigCommerce;
 
 class Plugin {
-	const VERSION = '0.13.0';
+	const VERSION = '0.14.0';
 
 	protected static $_instance;